# artifacts/vtvl-sim/src/lib/api.ts
/**
 * API client for the VTVL Simulator backend.
 *
 * TYPE CONTRACT — this project does not use OpenAPI or code-generated clients.
 * The interfaces defined in this file ARE the single source of truth for the
 * HTTP API contract between frontend and server. Any change to a server response
 * or request body shape MUST be reflected here.
 *
 * Key type distinctions:
 *  - RoomInfo           — shape returned by GET /rooms/:code (no hostSecret)
 *  - CreateRoomResponse — shape returned by POST /rooms (adds hostSecret)
 *  - SubmitRunPayload   — includes clientRunId for idempotent retries
 */
const API_BASE = "/api";

export interface UserInfo {
  id: string;
  username: string;
  email: string;
}

export interface RoomInfo {
  id: string;
  code: string;
  type: "coop" | "versus";
  missionId: string;
  status: string;
  hostId: string;
}

/**
 * Response from POST /rooms (room creation).
 * Extends RoomInfo with hostSecret — a server-issued credential returned
 * only to the creator and never exposed through GET /rooms/:code.
 * The creator must include this in the WebSocket join_room message to
 * claim host authority over the room.
 */
export interface CreateRoomResponse extends RoomInfo {
  hostSecret: string;
}

export interface LeaderboardEntry {
  id: string;
  displayName: string;
  userId?: string | null;
  missionId: string;
  score: number;
  grade: string;
  createdAt: string;
}

export interface PlayerStats {
  userId: string;
  username: string;
  totalFlights: number;
  bestScore: number | null;
  bestGrade: string | null;
  successRate: number;
}

export interface SubmitRunPayload {
  /**
   * Stable per-flight UUID generated by the client. The server uses this for
   * idempotency — retrying the same POST with the same clientRunId returns the
   * previously recorded result without inserting a duplicate row.
   */
  clientRunId?: string;
  missionId: string;
  guestName?: string;
  score: number;
  grade: string;
  crashed: boolean;
  touchdownSpeed?: number;
  padDeviation?: number;
  fuelRemaining?: number;
  tiltDeg?: number;
  flightDuration?: number;
}

async function apiFetch<T>(
  path: string,
  opts: RequestInit = {},
  token?: string,
): Promise<T> {
  const headers: Record<string, string> = {
    "Content-Type": "application/json",
  };
  if (token) headers["x-session-token"] = token;

  const res = await fetch(`${API_BASE}${path}`, {
    credentials: "include",
    headers,
    ...opts,
  });

  if (!res.ok) {
    const data = await res.json().catch(() => ({ error: res.statusText }));
    throw new Error(data?.error ?? `HTTP ${res.status}`);
  }
  return res.json() as Promise<T>;
}

export const api = {
  auth: {
    register: (username: string, email: string, password: string) =>
      apiFetch<{ token: string; user: UserInfo }>("/auth/register", {
        method: "POST",
        body: JSON.stringify({ username, email, password }),
      }),
    login: (email: string, password: string) =>
      apiFetch<{ token: string; user: UserInfo }>("/auth/login", {
        method: "POST",
        body: JSON.stringify({ email, password }),
      }),
    logout: (token?: string) =>
      apiFetch<{ ok: boolean }>("/auth/logout", { method: "POST" }, token),
    me: (token?: string) =>
      apiFetch<{ user: UserInfo | null }>("/auth/me", {}, token),
  },
  rooms: {
    create: (type: "coop" | "versus", missionId: string, guestName?: string, token?: string) =>
      apiFetch<CreateRoomResponse>(
        "/rooms",
        { method: "POST", body: JSON.stringify({ type, missionId, guestName }) },
        token,
      ),
    get: (code: string, token?: string) =>
      apiFetch<RoomInfo>(`/rooms/${code}`, {}, token),
  },
  runs: {
    submit: (payload: SubmitRunPayload, token?: string) =>
      apiFetch<{ id: string; leaderboardId: string | null }>(
        "/runs",
        { method: "POST", body: JSON.stringify(payload) },
        token,
      ),
  },
  leaderboard: {
    get: (missionId?: string, limit = 50, token?: string) => {
      const params = new URLSearchParams();
      if (missionId) params.set("missionId", missionId);
      params.set("limit", String(limit));
      return apiFetch<{ entries: LeaderboardEntry[] }>(`/leaderboard?${params}`, {}, token);
    },
  },
  players: {
    stats: (userId: string, token?: string) =>
      apiFetch<PlayerStats>(`/players/${userId}/stats`, {}, token),
  },
};