# lib/api-zod/src/generated/types/createRoomRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateRoomRequestType } from "./createRoomRequestType";

export interface CreateRoomRequest {
  type: CreateRoomRequestType;
  missionId: string;
  guestName?: string;
}