# lib/api-zod/src/generated/types/createRoomResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

import type { RoomResponse } from "./roomResponse";

/**
 * Returned only from POST /rooms. Extends RoomResponse with hostSecret,
 * a one-time credential the host must supply via the WS join_room message
 * to claim host authority. Never returned by GET /rooms/{code}.
 */
export interface CreateRoomResponse extends RoomResponse {
  /** UUID credential for the room host; never exposed to non-host clients. */
  hostSecret: string;
}